feat: settings card stories
