+++ shared/components/common/layout.tsx
@@ -2,10 +2,8 @@ import React from 'react';
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
   return (
-    <div className="bg-c_bg dark:bg-cd_bg h-[90vh]">
-      <div className="container p-8 bg-c_bg dark:bg-cd_bg mx-auto">
-        {children}
-      </div>
+    <div className="bg-c_bg dark:bg-cd_bg h-[90vh] relative p-4">
+      {children}
     </div>
   );
 };
